Progress on MollenOS
GDT and IDT is in place
Physical memory manager is now in place and tested. It works great. Used
memory so far is 600-700 kb, however most of it is memory reserved for
kernel (256 kb) and the bitmap. The used size will increase very slowly
from hereon out. The physical memory manager is protected by a spinlock.