VFS + MFS + MString
VFS:
I have defined structures and functions, and some enumerators with error
codes etc.

MFS:
Expanded with OpenFile and a core-meaty function that actually locates
files in the filesystem. So much more to do, and it makes heavy use of
the new string library (Known as MString)

MString:
Started builing a UTF-8 string library that will be used for MollenOS.
This will primarily be used in consumption with file-systems and paths.
For general output we still use ANSI.