A step towards a modular kernel
So for fun, I started building a module/driver interface and tried
porting my Usb stack + Ohci for a proof of concept.

But fuck, it's much work and I don't like it that much :O... Im thinking
of putting this step on hold and just finish my MFS and VFS
implementation. After that I could do it.