PE Loader
Finished most of the needed structures, I need to redo the Import
Header, it seems the one in the specification is not the one that is
used?!

Wrote most of the PE loader code, we can relocate the sections, rebase
the code (if there is relocation section) and load exported functions.

Next up is the import support in my PE Loader.