Modules
A lot of fixes on loading of modules. Now I have a completely new issue,
it SEEMS my kernel crashes with multiple accesses to my heap, which I
was convinced was smp-safe, but no, it is not.  To make completely sure,
i'll try to disable other cores and run, if it works, im correct.