Removed function table
So, made my kernel export functions rather than have a middleman
(function table) in function calls across modules. We also modified the
bootloader to support this and improved the way information is passed to
the kernel.

I also implemented Critical Sections to use in the memory systems &
fixed a deadlock. Buuuuut now it seems i just get an instant triple
fault on real hardware during the UsbCore setup. For what i debugged
it's related to an interrupt (OHCI) while the UsbCore is running it's
port setup