A lot of fixes
MFS:
Fixes to the calculation of the bucket map.
Fixes to GetNextBucket/SetNextBucket
Fixes to ReadFile

Vfs:
Fixed the canonicalize path.

PeLoader/ProcessManager:
Fixed loading of libraries.

Next is the fault in my port of dlmalloc, it faults on a NULL pointer
somewhere. Probably due to my memory allocation in userspace. Gotta
inspect the values