Restructered modules
So i restructured all my modules so they fit better. Also I've begun to
work on a bus driver and I'm refactoring my device-system so it's more
flexible and to support more driver-independance.

Currently I'm stuck on the bus driver and how it should work / interact
with the rest of the system.