Fixed spinlock export + RTC/PS2/PIT
Fixed an issue with kernel not exporting spinlockreset, and finished the
rewrite of following drivers:
RTC,
PS2,
PIT.

This leaves all the "shared" architecture drivers, which is not really
that many for now. I'm doing HPET first, because then I can test my
whole new system. I'm not enabling the bus-driver before it works