Threading/Scheduler rewrite, UsbHid fix, pci enumeration fix
So OUT OF THE BLUE a scheduler error appeared, and it was a sort of "why
does nothing work now" bug. So I decided to review my whole threading
and scheduler and it dawned upon me ... "What the FRICK did I do to my
scheduler?!" Anyway rewrote much of the system to work with seperate
lists, seperate systems instead of being interlinked as they were.
Everything works much better now.

The scheduler has a weakness though, it needs new list-nodes EACH time
it schedules a task, which is serious shit, to counter-act this problem
I made a node-recycler in the scheduler, but at some point this probably
will have to be fixed for performance.