EHCI - Fixes
So, accidently allocated the async head (forgotten to set it
preallocated).. Caused a freakin mess of issues. Fixed some wake-code in
EHCI INTR handler. Now bulk and control work like a charm except for ONE
case. When I try a bulk transfer (IN --> 0x1000, MPS 0x200), it never
seems to ... do it. First transfer that is beyond the MPS.