Fixed seek/read
VFS:
Had to disable input buffering, as the current implementation was really
bad it seemed. Now the output buffering is working pretty well, and I
actually think I'm going to implement the input buffering in userspace
to speed it up as well.

MFS:
Both reading and writing is SLOW, but works. This is mainly due to the
reason that if the file is allocated into ONE HUGE bucket, then it loads
the entire file every time it just wants to read a single byte. So i
have to calculate which bucket I need to read before hand instead, so I
at max load only a bucket, and not the entire file