Updates to CLIB, bugfixes, updated tools
Updated a lot of userspace to new builds. I had made a lot of changes to
the c-library and thus everything needed a rebuild. I have encounted a
few new problems tho.

Bootloader has been updated and the VESA mode selector has been fixed.
It seemed it only worked when the returned list of supported modes were
in the first 64 kb, otherwise the segment was not selected.

MfsTool has been updated to support creation of disk images as well as
real harddrives, running MfsTool with -vmdk now creates a vmdk image
that actually works (tested!), two new build scripts has been added to
reflect this.

Fixed a null pointer in UHCI, but the scheduling code needs to be looked
at to make sure everything is ok. I also fixed an error in the interrupt
code that could override the InterruptLine of a MCoreDevice. I found
some other errors when spawning processes etc. And i really need to fix
why my OS just crashes on VMWare