Everything works, except for loading of large png files
Loading large png files (> 2mb atleast), it still gives the "Bad
Adaptive Filter Value" error, which indicate the data is corrupt. Now it
doesn't exactly seem like this issue is due to my libc anymore, I'm
starting to think this issue is actually because of the memory
allocators OR the data-transfers.