Work on #136
It doesn't matter how much data is asked. But i can derive from the
bochs log it doesn't queue up the packets that it defer, it seems it
only saves the last packet and ignores the rest which is insanely wierd.
They are queued up on the msd part of bochs, but they are incorrectly
queued. I think this might be a bochs issue.